Fixed savings account creation

+++ src/main/java/com/dollarsbank/model/SavingsAccount.java
@@ -9,7 +9,7 @@ public class SavingsAccount {
 	private double funds;
 
 	public void setAccountNumber(int account_number) { this.account_number = account_number; }
-	public double getAccountNumber() { return this.account_number; }
+	public int getAccountNumber() { return this.account_number; }
 
 	public void setCustomer(Customer customer) { this.customer = customer; }
 	public Customer getCustomer() { return this.customer; }
